Make the java not to force close when sharedDb is specified
So, the latest release `v0.4.0` have a bug. If one specifies the `sharedDb: true` then local-dynamo cannot be started. It prints:
```
Unrecognized option: -sharedDb
Error: Could not create the Java Virtual Machine.
Error: A fatal exception has occurred. Program will exit.
```

This PR fixes it all! We just need to move the `-sharedDb` switch to the end of the command line.

<img src="https://i.makeagif.com/media/6-17-2015/wMDHKb.gif" />